fix: テストモックの修正と動的TTL戦略の準備
- popular-tags-display.test.tsxのモックを修正
  - requestThrottleとrankingCacheの内部Map構造を含む完全なモック
  - APIエンドポイントのパスを/api/edge/rankingから/api/rankingに統一
  - popular-tags APIのモックを追加
  - filterWithNGListとmigrateLocalStorageDataのモックを追加
  - DecompressionStreamのモックを追加（ブラウザAPI互換性対応）

- API routeの開発環境対応を改善
  - route.tsとroute.dev.tsに分離して可読性を向上
  - route.prod.tsを削除（route.tsで統合）
  - console.logをコメントアウト（ESLint対応）

- wrangler-prod.tomlを追加
  - 本番環境用のWorker設定ファイル
  - 動的TTL戦略の実装に向けた準備
