+++ app/api/ranking/route.ts
@@ -1,113 +1,170 @@
 import { NextRequest, NextResponse } from 'next/server'
+import { getGenreRanking, getTagRanking } from '@/lib/cloudflare-kv'
+import { generateMockRankingData, generateMockPopularTags, isDevelopmentWithoutKV } from '@/lib/mock-data'
+import type { RankingGenre, RankingPeriod } from '@/types/ranking-config'
+import type { RankingItem } from '@/types/ranking'
 
-// 開発環境用: 外部Workerにプロキシする設定
+// Node.js Runtimeを使用（環境変数の問題を回避）
 export const runtime = 'nodejs'
 
 export async function GET(request: NextRequest) {
   const { searchParams } = new URL(request.url)
+  const genre = searchParams.get('genre') || 'all'
+  const period = searchParams.get('period') || '24h'
+  const tag = searchParams.get('tag') || undefined
+
+  // Validate inputs - period のみチェック（genreはすべて受け入れる）
+  const validPeriods = ['24h', 'hour']
   
-  // 環境変数から開発Worker URLを取得
-  const devWorkerUrl = process.env.NEXT_PUBLIC_API_GATEWAY_URL || 
-                       'https://nico-ranking-dynamic-cache-dev.yjsn180180.workers.dev'
-  
-  // プロキシ先URLを構築
-  const proxyUrl = new URL('/api/ranking', devWorkerUrl)
-  proxyUrl.search = searchParams.toString()
-  
+  if (!validPeriods.includes(period)) {
+    return NextResponse.json({ error: 'Invalid period' }, { status: 400 })
+  }
+
   try {
-    console.log(`[DEV] Proxying to: ${proxyUrl.toString()}`)
-    
-    // リクエストヘッダーを転送（If-None-Matchなど）
-    const requestHeaders: HeadersInit = {
-      'Accept': 'application/json',
-      'Accept-Encoding': 'gzip, deflate, br',
-      'User-Agent': 'NextJS-Dev-Proxy/1.0'
-    }
-    
-    // If-None-Matchヘッダーがある場合は転送
-    const ifNoneMatch = request.headers.get('if-none-match')
-    if (ifNoneMatch) {
-      requestHeaders['If-None-Match'] = ifNoneMatch
-    }
-    
-    const response = await fetch(proxyUrl.toString(), {
-      method: 'GET',
-      headers: requestHeaders
-    })
-    
-    // 304 Not Modifiedの場合は特別な処理
-    if (response.status === 304) {
-      const responseHeaders = new Headers()
+    // 開発環境でKVが設定されていない場合はモックデータを返す
+    if (isDevelopmentWithoutKV()) {
+      // eslint-disable-next-line no-console
+      console.log('[Development] Using mock data - KV not configured')
       
-      // 必要なヘッダーをコピー
-      response.headers.forEach((value, key) => {
-        const lowerKey = key.toLowerCase()
-        if (lowerKey.startsWith('x-') || 
-            lowerKey === 'cache-control' ||
-            lowerKey === 'cdn-cache-control' ||
-            lowerKey === 'etag') {
-          responseHeaders.set(key, value)
+      const mockItems = generateMockRankingData(500)
+      const mockPopularTagObjects = generateMockPopularTags()
+      // PopularTagオブジェクトから名前のみの配列に変換
+      const mockPopularTags = mockPopularTagObjects.map(tag => tag.name)
+      
+      // タグ別ランキングの処理
+      if (tag) {
+        // タグに該当する動画のみフィルタリング
+        const taggedItems = mockItems
+          .filter(item => item.tags.includes(tag))
+          .slice(0, 100) // タグ別は最大100件
+        
+        if (taggedItems.length === 0) {
+          return NextResponse.json(
+            { 
+              error: 'Tag ranking not found. This tag may not be in the popular tags list.',
+              items: [],
+              hasMore: false,
+              totalCached: 0
+            },
+            { status: 404 }
+          )
         }
-      })
-      
-      // 開発環境識別用ヘッダー追加
-      responseHeaders.set('X-Proxy-Source', 'nextjs-dev')
-      responseHeaders.set('X-Worker-URL', devWorkerUrl)
-      
-      return new Response(null, {
-        status: 304,
-        headers: responseHeaders
-      })
-    }
-    
-    if (!response.ok) {
-      throw new Error(`Worker response: ${response.status} ${response.statusText}`)
-    }
-    
-    // レスポンスボディを処理（圧縮も考慮）
-    const contentEncoding = response.headers.get('content-encoding')
-    let data: ArrayBuffer | string
-    
-    if (contentEncoding && contentEncoding.includes('gzip')) {
-      // 圧縮されたデータはそのまま転送
-      data = await response.arrayBuffer()
-    } else {
-      data = await response.text()
-    }
-    
-    const responseHeaders = new Headers()
-    
-    // 必要なヘッダーをコピー
-    response.headers.forEach((value, key) => {
-      const lowerKey = key.toLowerCase()
-      if (lowerKey.startsWith('x-') || 
-          lowerKey === 'cache-control' ||
-          lowerKey === 'cdn-cache-control' ||
-          lowerKey === 'etag' ||
-          lowerKey === 'content-encoding') {
-        responseHeaders.set(key, value)
+        
+        const response = NextResponse.json({
+          items: taggedItems,
+          hasMore: false,
+          totalCached: taggedItems.length
+        })
+        response.headers.set('X-Data-Source', 'mock')
+        response.headers.set('X-Environment', 'development')
+        return response
       }
-    })
+      
+      // ジャンル別ランキング（開発環境ではジャンルフィルタリングなし）
+      const response = NextResponse.json({
+        items: mockItems,
+        popularTags: mockPopularTags,
+        hasMore: false,
+        totalCached: mockItems.length
+      })
+      response.headers.set('X-Data-Source', 'mock')
+      response.headers.set('X-Environment', 'development')
+      return response
+    }
     
-    // 開発環境識別用ヘッダー追加
-    responseHeaders.set('X-Proxy-Source', 'nextjs-dev')
-    responseHeaders.set('X-Worker-URL', devWorkerUrl)
-    responseHeaders.set('Content-Type', 'application/json')
+    // Cloudflare KVが利用可能かチェック（環境変数で判定）
+    const kvRankingId = process.env.KV_RANKING_ID?.trim()
+    const cloudflareApiToken = process.env.CLOUDFLARE_API_TOKEN?.trim()
+    const useCloudflareKV = kvRankingId && cloudflareApiToken && kvRankingId !== '' && cloudflareApiToken !== ''
     
-    return new Response(data, {
-      status: response.status,
-      headers: responseHeaders
-    })
+    // CI環境でのデバッグ情報
+    if (process.env.CI) {
+      // eslint-disable-next-line no-console
+      console.log('CI Environment Debug:', {
+        hasKvRankingId: !!kvRankingId,
+        hasApiToken: !!cloudflareApiToken,
+        kvIdLength: kvRankingId?.length || 0,
+        apiTokenLength: cloudflareApiToken?.length || 0,
+        useCloudflareKV
+      })
+    }
+    
+    // タグ別ランキングの処理
+    if (tag) {
+      // Cloudflare KVからの取得を試みる
+      if (useCloudflareKV) {
+        try {
+          const cfItems = await getTagRanking(genre, period as RankingPeriod, tag)
+          if (cfItems && cfItems.length > 0) {
+            // タグ別ランキングは全件返す（KVに保存されている分すべて）
+            const response = NextResponse.json({
+              items: cfItems, // 全件返す（239件など）
+              hasMore: false, // タグ別ランキングは常にfalse
+              totalCached: cfItems.length
+            })
+            response.headers.set('Cache-Control', 'public, s-maxage=3600, stale-while-revalidate=1800')
+            response.headers.set('X-Cache-Status', 'CF-HIT')
+            response.headers.set('X-Total-Cached', cfItems.length.toString())
+            response.headers.set('X-API-Version', '2') // バージョン確認用
+            return response
+          }
+        } catch (error) {
+          // Cloudflare KV error - silently fallback to dynamic fetch
+        }
+      }
+      
+      // KVにデータがない場合は、タグが人気タグリストにない可能性がある
+      // 404 Not Foundを返す（500エラーではなく）
+      return NextResponse.json(
+        { 
+          error: 'Tag ranking not found. This tag may not be in the popular tags list.',
+          items: [],
+          hasMore: false,
+          totalCached: 0
+        },
+        { status: 404 }
+      )
+    }
+
+    // 通常のジャンル別ランキング
+    
+    // Cloudflare KVからの取得を試みる
+    if (useCloudflareKV) {
+      try {
+        const cfData = await getGenreRanking(genre, period as RankingPeriod)
+        if (cfData && cfData.items && cfData.items.length > 0) {
+          // ジャンル別ランキングは500件まで返す
+          const maxItems = 500
+          const items = cfData.items.slice(0, maxItems)
+          
+          const response = NextResponse.json({
+            items: items,
+            popularTags: cfData.popularTags || [],
+            hasMore: false, // ページネーションなし
+            totalCached: cfData.items.length
+          })
+          response.headers.set('Cache-Control', 'public, s-maxage=3600, stale-while-revalidate=1800')
+          response.headers.set('X-Cache-Status', 'CF-HIT')
+          response.headers.set('X-Max-Items', String(maxItems))
+          response.headers.set('X-API-Version', '2') // バージョン確認用
+          return response
+        }
+      } catch (error) {
+        // Cloudflare KV error - silently fallback to dynamic fetch
+      }
+    }
+    
+    // KVにデータがない場合はエラーを返す
+    return NextResponse.json(
+      { error: 'ランキングデータが見つかりません。しばらくしてから再度お試しください。' },
+      { status: 503 }
+    )
     
   } catch (error) {
-    console.error('[DEV] Proxy error:', error)
+    // API error - return error response
     return NextResponse.json(
-      { 
-        error: 'Development Worker connection failed',
-        workerUrl: devWorkerUrl,
-        details: error instanceof Error ? error.message : String(error)
-      },
-      { status: 502 }
+      { error: 'ランキングデータの取得に失敗しました。' },
+      { status: 503 }
     )
   }
 }
